fix(build): eliminate all int-to-pointer-cast warnings (284→0, x86_64 native)
- Add uintptr/uintptr_t intermediate casts in REG_READ32/16/8 macros
  (Compiler.h, Std_Types.h, Mcal.h, Icu_Private.h)
- Fix inline pointer casts in MCAL/ECUAL: RamTst, RamSafety, Flash, Fls,
  Fls_Hw, Eep, Icu_Irq, Mcu, Dio, Adc, Port, Wdg, Pwm, Gpt, Can, I2c, Xcp,
  Boot_Loader (AppEntry function pointer), Platform_EccHandler
- Native x86_64 build: 0 errors, 0 warnings (was 284 int-to-pointer-cast)
- Embedded ARM target behavior unchanged (uintptr is 32-bit there)
- Also include pending DoIP/SecOC fixes from previous round

+++ include/autosar/Compiler.h
@@ -62,16 +62,16 @@
 
 /* Memory-mapped register I/O macros (native stub - real HW access) */
 #ifndef REG_READ32
-#define REG_READ32(addr)                (*((volatile uint32*)(addr)))
+#define REG_READ32(addr)                (*((volatile uint32*)(uintptr)(addr)))
 #endif
 #ifndef REG_WRITE32
-#define REG_WRITE32(addr, val)          (*((volatile uint32*)(addr)) = (val))
+#define REG_WRITE32(addr, val)          (*((volatile uint32*)(uintptr)(addr)) = (val))
 #endif
 #ifndef REG_READ8
-#define REG_READ8(addr)                 (*((volatile uint8*)(addr)))
+#define REG_READ8(addr)                 (*((volatile uint8*)(uintptr)(addr)))
 #endif
 #ifndef REG_WRITE8
-#define REG_WRITE8(addr, val)           (*((volatile uint8*)(addr)) = (val))
+#define REG_WRITE8(addr, val)           (*((volatile uint8*)(uintptr)(addr)) = (val))
 #endif
 
 #endif /* COMPILER_H */
